Refactor: cli.py: Move config handling to main
Remove the global config handling as well as the global variables. Move
the handling of the config to the main function and remove the
assignment and creation function as the amount of configuration
attributes is small and the structure is quite simple.
Remove the config package as it is not required anymore.
Insert the configuration as parameter into the Windows class.

Signed-off-by: Sebastian Fricke <sebastian.fricke@posteo.net>